board: Add support for Nypon board
This is basically a copy and edit from the stemmy board but
the changes that were needed required a new board as the SONY
bootloader disabled the clocks for a lot of stuff. This means
with a code from minecrell the clocks are up in this board and
eMMC works. This is the only thing that works and was tested tho